requirements.txt
openai>=1.30.0
python-dotenv>=1.0.1
requests>=2.32.3
beautifulsoup4>=4.12.3
langchain>=0.2.0
langchain-openai>=0.1.8
faiss-cpu>=1.8.0
numpy>=1.26.0
